pyproject.toml
[tool.poetry]
name = "travel-agency"
version = "0.1.0"
description = ""
authors = ["nariman079 <nariman079i@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
postgres = "^4.0"
celery = "^5.4.0"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
